Add start meny
Added a start meny where you can choose between three starters and see both a sprite of the starter and some stats. I also fixed the wild grass so you dont need to have a velocity that is not equal to zero in both the y and x axis. Aka just one axis is enough now.